Moved all uniform distribution related code into its own file
